feat(linkPreview): add image resize, transform to mutation, convert to webp
