+++ api/hub/package.json
@@ -44,7 +44,6 @@
     "apollo-server-cache-redis": "3.1.0",
     "apollo-server-koa": "3.3.0",
     "apollo-server-plugin-base": "3.2.0",
-    "cross-fetch": "3.1.4",
     "dotenv": "10.0.0",
     "emittery": "0.10.0",
     "eth-ens-namehash": "2.0.8",
@@ -63,9 +62,11 @@
     "link-preview-js": "2.1.8",
     "lru-cache": "6.0.0",
     "mailgun-js": "^0.22.0",
+    "node-fetch": "2.6.5",
     "node-abort-controller": "3.0.0",
     "object-hash": "2.2.0",
     "prom-client": "13.2.0",
+    "sharp": "0.29.1",
     "winston": "3.3.3"
   },
   "devDependencies": {
