+++ sdk/src/posts/entry.graphql
@@ -124,7 +124,7 @@ fragment LinkPreviewF on LinkPreview {
   images
 }
 
-query GetLinkPreview($link: String!) {
+mutation GetLinkPreview($link: String!) {
   getLinkPreview(link: $link){
     ...LinkPreviewF
   }
